FP-19: Remove isAuthenticatorRegistered api
This method is not actually callable on iOS and shouldn't be handled by
the sdk in the first place as you can retrieve a list of all registered
authenticators. So let's remove this from the api.

There are remnants of this still in the kotlin/swift code, we will clean
that up in a later commit.