FP-19: Create wrapperTypeError, catch typeErrors, throw platformExc
It is possible for the native sdk to return a different type then the
dart side expects. We can naively catch this by catching TypeErrors and
then throwing a PlatformException. This PlatformException uses the
code and message of the wrapperTypeError.

This is not an optimal solution but it does help with nullSafety. This
will be replaced in a later commit with the use of a library called
Pigeon which will guarantee type safety for the native functions.