FP-49: Rework a large part of pin handling
Previously there was a lot of incoherent handling of the pins,
especially for changing pin flow. This commit updates the change pin
behavior to use the already existing login/register behaviors. It also
removes a lot of initalizers in the BridgeConnector and various pin
handling classes.

It also updates the dart api to actually use the pigeon api instead of
the message channels for starting the changepin flow.