FP-71: Move startApp to pigeon for iOS/Android
This commit updates startApp to go through pigeon, for this we need to
update a few types that get passed along and need to update how we
create some of the main objects on the dart side. e.g. the native
interface api.
It also has some changes on how we deal with the config on the Android
side.