FP-57: Don't expose the pin cr. callback and throw correct error when not in progress
Previously we exposed the callback through a companion object. This is
not desirable as anything can happen to that object outside of the
class. This commit moves all the logic into the handler class and
exposes two public methods to interact with the callback which will
throw errors if the callback is null.

This commit also updates the error thrown during pin creation as pin
creation should not give the error for registration_not_in_progress as
it's something else.