FP-32: Add .swiftlint.yml to exclude tests and include the devpod
Flutter projects get set up by default without being able to compile the
plugin by itself. It will add all the code as a development pod so we
will need to add this one a bit hackish through the symlink that it
creates.

We also exclude the tests for now.