FP-81: Create new delegate on authenticateUser calls
Previously we were overwriting the stored callback on calling a method
twice in a row. This would result in the answer back to Flutter never
occuring because it would try to call the last completion twice.

This would result in the Future in Flutter never resolving.

This commit fixes that by creating seperate delegate objects for every
call. These are handled correctly by the underlying native sdk so we
don't have to worry about that.