FP-81: Create new delegate object with promise completion on calling register
We create a new delegate object when calling registerUser with the
Future completion. This makes sure that when we call registerUser twice
we don't break our Futures.

Note: We had some code which would make the Future fail if somehow the
browserRegistrationChallenge was gone when the internal browser would
either cancel or succeed.
This has been removed and we will have to make sure that we never set
this object to nil except when we responded to the Future.
This implementation is not perfect, but it is clearly better and can be
improved later when we don't have to handle the browser inside the sdk.